requirements.txt
numpy==1.20.1
pytest==6.2.2
colour==0.1.5
Pillow==8.1.2
